Backend/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
arrow==1.2.3
astroid==2.12.12
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
dill==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.2
fastapi==0.87.0
flake8==5.0.4
Flask==2.2.2
frozenlist==1.3.1
greenlet==2.0.1
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
motor==3.0.0
multidict==6.0.2
odmantic==0.9.0
orjson==3.8.1
packaging==21.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyflakes==2.5.0
PyJWT==2.6.0
pylint==2.15.5
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
starlette==0.21.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.19.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
yarl==1.8.1
